+++ middlewares/osio/jwt_token.go
@@ -1,4 +1,4 @@
-package middlewares
+package osio
 
 import (
 	"crypto/rsa"
