+++ middlewares/osio/osio.go
@@ -1,4 +1,4 @@
-package middlewares
+package osio
 
 import (
 	"crypto/sha256"
