+++ middlewares/osio/osio_basic_test.go
@@ -1,4 +1,4 @@
-package middlewares
+package osio
 
 import (
 	"fmt"
