+++ middlewares/osio/osio_che_test.go
@@ -1,4 +1,4 @@
-package middlewares
+package osio
 
 import (
 	"fmt"
