+++ middlewares/osio/tenant.go
@@ -1,4 +1,4 @@
-package middlewares
+package osio
 
 import (
 	"encoding/json"
